Added better class names to Database.
