+++ .gitignore
@@ -1,3 +1,4 @@
+__pycache__/
 No-Instructor-data.txt
 # .idea
 .vscode
